migrate from Adafruit_SSD1306 to SSD1306Ascii
The Adafruit_SSD1306 is using a full page in SRAM
memory with a size of approx 1.5KB, taking approx
75% of the SRAM memory. The usage of the library
causes the memory issues and prevents from further
extending the logic within this application

The SSD1306Ascii has a lot less functionality but
it's taking only 100bytes of the SRAM memory
which leaves a lot more to use with the actual
program. The downside is that it's less optimized
and has a slower speed. As a result the screen
flickers when it clears / draws.
It's probably possible by replacing the complete
Clear of the screen by only clearing sections which
are required to be cleared.
Maybe in the future, for now it serves its purpose.

The chart logic is disabled as it requires to be
redefined.