load and save profiles to EEPROM
implementation of logic to load
and save profiles to the EEPROM.
Each profile exists out of 11 bytes
Based on the number of bytes / variables
within the profile class, the position in
the eeprom is defined. At the moment only 3
profiles are supported. This may get expanded
or even be made dynamic.