refactor ReflowController
The controller needs to be adjusted to work
optimally with the current design. This commit
entails a part of the work done; more commits to
come.

Reflow controller accepts MenuSystem so that we can
use it to write to the display. Display requires
to be updated with additional logic for notifications
and graphs.

MAX6675 is the selected input device for the thermocouple
and proviced temperature information for the reflow process.
Might just initialize it as part of the controller instead
of in the main loop as it's unused there.

Profile provides information on the configured profile. The
profile can still be expanded with additional settings, like
the PID controller settings and it needs to be linked to the
process.

Encoder is used to verify if there is any input. It's passed on here
as the reflow process runs in a dedicated while(true) loop. Might move
it to the menuSystem object in the future.

heatpin1 and heatpin2 are the output pins for the SSR controlling the
heaters.